+++ src/NServiceBus.WindowsPerformanceCounters/NServiceBus.WindowsPerformanceCounters.csproj
@@ -56,7 +56,9 @@
     <Compile Include="Counters\PerformanceMonitorUsersInstaller.cs" />
     <Compile Include="CriticalTime\CriticalTimeMonitoring.cs" />
     <Compile Include="Metrics.cs" />
-    <Compile Include="PerfCounters.cs" />
+    <Compile Include="PerformanceCounters.cs" />
+    <Compile Include="PerformanceCountersEndpointConfigurationExtensions.cs" />
+    <Compile Include="PerformanceSettings.cs" />
     <Compile Include="Performance\ReceiveStatisticsConfig.cs" />
     <Compile Include="Performance\ReceiveStatisticsFeature.cs" />
     <Compile Include="Performance\Statistics\ReceivePerformanceDiagnosticsBehavior.cs" />
